# fix sigmoid