Simpler, more accurate implementation of sigmoid.

Fix https://github.com/FluxML/NNlib.jl/issues/78

I benchmarked this and there is no performance difference.

After this PR:

```
julia> ForwardDiff.derivative(σ, -710)
4.47628622567513e-309
julia> x = 7.0; @btime σ($x);
  12.757 ns (0 allocations: 0 bytes)
julia> x = -7.0; @btime σ($x);
  13.704 ns (0 allocations: 0 bytes)
```

Before this PR:

```
julia> ForwardDiff.derivative(σ2, -710)
NaN
julia> x = 7.0; @btime σ($x);
  13.906 ns (0 allocations: 0 bytes)
julia> x = -7.0; @btime σ($x);
  13.366 ns (0 allocations: 0 bytes)
```